feat: multiplex

+++ app/search_engine/repository/db/dao/input_data.go
@@ -35,7 +35,7 @@ func (d *InputDataDao) ListInputData() (in []*model.InputData, err error) {
 }
 
 // ListInputDataByDocIds 根据传进来的 docs id 获取所有的信息
-func (d *InputDataDao) ListInputDataByDocIds(docIds []uint32) (in []*types.SearchItem, err error) {
+func (d *InputDataDao) ListInputDataByDocIds(docIds []int64) (in []*types.SearchItem, err error) {
 	err = d.DB.Model(&model.InputData{}).
 		Where("doc_id IN ?", docIds).
 		Select("doc_id," +
